Add XGBoost artifacts; run integration tests inside Jenkins
