+++ tests/python/integration/xgboost/iris.libsvm
@@ -1,0 +1,1 @@
+1 0:6.1 1:2.8 2:4.7 3:1.2
