+++ tests/python/integration/xgboost/mnist.libsvm
